Adding feedback changes for helper functions
